solutions/django/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-bootstrap3==14.1.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
idna==2.10
importlib-metadata==1.7.0
phonenumbers==8.12.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.2.0
gunicorn==20.0.4